Refactor DMARC
Deduplicate TXT records for DMARC and use native builder for better
formatting. Move "spf" include to a more generic "email" one to avoid an
additional file only for a single DMARC builder block. Add new DMARC
macro to defaults as it makes sense to have the same for all zones.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>