Allow different defaults for reverse zones
These have a different TTL and do not need email related records.
Split our defaults into a common set and a smaller one for use with
reverse zones. This requires setting DEFAULTS(DEFAULT_BASE) at the
beginning of affected zones and DEFAULTS(DEFAULT_ALL) (to reset to the
original defaults for later includes) at the end - no cleaner way was
found so far.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>