Restructure
The single directory tended to cause dependency issues, if referenced
functions were declared in a file with a later include order. Instead of
more awkward numbering and underscoring, keep the "zones" directory for
files which really define zones and move the helper logic and variables
to a separate one. This slightly complicates the init logic, as recursive
includes cannot be used for the new "common" directory, but this is deemed
a reasonable compromise as we do not introduce new files there as often.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>